fix require stucture
- goal: files like daru/dataframe should be funtional on their own
- daru/dataframe could not be required because it calls Daru.has_nyaplot?
- Daru.has_nyaplot? is defined in lib/daru.rb, which pulls in all of daru
  including dara/dataframe
- to avoid circular dependencies and to allow files to stand on their own, it
  is necessary to separate critical functionality from require-the-world
- move critical, common functionality like Daru.has_nyaplot? into
  daru/platform
- require daru/platform where needed, like lib/daru and daru/dataframe
- now, this works: require 'daru/dataframe'
